Turned it into a WORKING maven project
pretty sicccccccc
