+++ pom.xml
@@ -1,0 +1,66 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>SkribbleHax</groupId>
+  <artifactId>SkribbleHax</artifactId>
+  <version>0.1</version>
+  <name>SkribbleHax</name>
+<dependencies>
+	<dependency>
+	    <groupId>com.jgoodies</groupId>
+	    <artifactId>forms</artifactId>
+	    <version>1.2.1</version>
+	</dependency>
+	<dependency>
+    	<groupId>com.miglayout</groupId>
+    	<artifactId>miglayout</artifactId>
+    	<version>3.7.4</version>
+	</dependency>
+    <dependency>
+		<groupId>junit</groupId>
+		<artifactId>junit</artifactId>
+		<version>3.8.1</version>
+		<scope>test</scope>
+    </dependency>
+    <dependency> <!-- JNA dependency -->
+		<groupId>net.java.dev.jna</groupId>
+		<artifactId>jna</artifactId>
+		<version>4.5.2</version>
+    </dependency>
+    <dependency> <!-- JNA dependency -->
+		<groupId>net.java.dev.jna</groupId>
+		<artifactId>jna-platform</artifactId>
+		<version>4.5.2</version>
+    </dependency>
+  </dependencies>
+  <properties>
+    <maven.compiler.source>1.8</maven.compiler.source>
+    <maven.compiler.target>1.8</maven.compiler.target>
+  </properties>
+  <build>
+  <plugins>
+    <plugin>
+      <groupId>org.apache.maven.plugins</groupId>
+      <artifactId>maven-assembly-plugin</artifactId>
+      <executions>
+        <execution>
+          <goals>
+            <goal>attached</goal>
+          </goals>
+          <phase>package</phase>
+          <configuration>
+            <descriptorRefs>
+              <descriptorRef>jar-with-dependencies</descriptorRef>
+            </descriptorRefs>
+            <archive>
+              <manifest>
+                <mainClass>skribbleHax.GUI</mainClass>
+              </manifest>
+            </archive>
+          </configuration>
+        </execution>
+      </executions>
+    </plugin>
+  </plugins>
+  </build>
+  
+</project>
